+++ integroBridgeService/src/main/java/com/adobe/campaign/tests/bridge/service/IntegroAPI.java
@@ -32,6 +32,7 @@ public class IntegroAPI {
     protected static final String ERROR_IBS_RUNTIME = "Problems with payload.";
     protected static final String ERROR_AMBIGUOUS_METHOD = "No unique method could be identified that matches your request.";
     private static final Logger log = LogManager.getLogger();
+    protected static final String ERROR_JAVA_OBJECT_NOT_ACCESSIBLE = "The java object you want to call is inaccessible. This is very possibly a scope problem.";
 
     public static void startServices(int port) {
 
@@ -94,7 +95,7 @@ public class IntegroAPI {
             res.status(statusCode);
             res.type(ERROR_CONTENT_TYPE);
             res.body(BridgeServiceFactory.createExceptionPayLoad(
-                    new ErrorObject(e, ERROR_AMBIGUOUS_METHOD, statusCode)));
+                    new ErrorObject(e, ERROR_AMBIGUOUS_METHOD, statusCode,false)));
         });
 
         exception(IBSConfigurationException.class, (e, req, res) -> {
@@ -124,14 +125,22 @@ public class IntegroAPI {
             res.status(statusCode);
             res.type(ERROR_CONTENT_TYPE);
             res.body(BridgeServiceFactory.createExceptionPayLoad(
-                    new ErrorObject(e, ERROR_JAVA_OBJECT_NOT_FOUND, statusCode)));
+                    new ErrorObject(e, ERROR_JAVA_OBJECT_NOT_FOUND, statusCode, false)));
+        });
+
+        exception(JavaObjectInaccessibleException.class, (e, req, res) -> {
+            int statusCode = 404;
+            res.status(statusCode);
+            res.type(ERROR_CONTENT_TYPE);
+            res.body(BridgeServiceFactory.createExceptionPayLoad(
+                    new ErrorObject(e, ERROR_JAVA_OBJECT_NOT_ACCESSIBLE, statusCode, false)));
         });
 
         exception(IBSTimeOutException.class, (e, req, res) -> {
             int statusCode = 408;
             res.status(statusCode);
             res.type(ERROR_CONTENT_TYPE);
-            res.body(BridgeServiceFactory.createExceptionPayLoad(new ErrorObject(e, ERROR_CALL_TIMEOUT, statusCode)));
+            res.body(BridgeServiceFactory.createExceptionPayLoad(new ErrorObject(e, ERROR_CALL_TIMEOUT, statusCode, false)));
         });
 
         //Internal exception